move generators so they can be found by rails
