+++ lib/active_model_serializers/railtie.rb
@@ -26,7 +26,7 @@ module ActiveModelSerializers
     end
 
     generators do
-      require 'generators/serializer/resource_override'
+      require 'generators/rails/resource_override'
     end
   end
 end
