+++ lib/generators/rails/resource_override.rb
